Fix #34 foreground color logic
This commit also removes optimization that stopped GPT timer when
reactive profiles had no LEDs enabled. That logic was too hard to
maintain. In addition, a new handler is added for explicitly clearing
the foreground color. Previously setProfile handler was reused for
that and it was hard to handle different edge cases arising from
different user setups.