Limit color bleed.
On reactive profiles the bleed might be gone (when the previous column is not
lit). On dimmer settings it should be much smaller, and in general it will be a
bit smaller because of pwmLimit exceeding the max possible color.

This might (?) limit power consumption, especially on reactive profiles, as
columns won't be enabled without a need.